LCD-14591 copy all files from src/**/* to static/
