+++ modules/sdk/gradle-plugins-workspace/src/gradleTest/clientExtensionSamples/sample-theme-css/src/css/_custom.scss
@@ -1,1 +1,5 @@
 @import 'sassy-inputs/sass/main';
+
+body {
+	background-image: url('img/paper.gif');
+}
